# Conflicts:
#	app/BUILD/src/DeskPRO/Bundle/AppBundle/Resources/config/services/features.yml
#	app/BUILD/src/DeskPRO/Bundle/UpdateBundle/BuildTasks/build-manifest.php

+++ app/BUILD/src/Application/LegacyApiBundle/Controller/CustomFieldsController.php
@@ -10,6 +10,12 @@ use Application\DeskPRO\CustomFields\FieldDisplayArray;
 use Application\DeskPRO\CustomFields\FieldManager;
 use Application\DeskPRO\CustomFields\Handler\Choice;
 use Application\DeskPRO\CustomFields\Handler\HandlerAbstract;
+use Application\DeskPRO\Entity\CustomDefArticle;
+use Application\DeskPRO\Entity\CustomDefChat;
+use Application\DeskPRO\Entity\CustomDefDownload;
+use Application\DeskPRO\Entity\CustomDefOrganization;
+use Application\DeskPRO\Entity\CustomDefPerson;
+use Application\DeskPRO\Entity\CustomDefTicket;
 use Application\DeskPRO\Entity\CustomFieldDefinition;
 use Application\DeskPRO\Entity\Ticket;
 use Application\DeskPRO\EntityRepository\CustomDefAbstract;
@@ -279,12 +285,12 @@ class CustomFieldsController extends AbstractController implements ProtectedCont
     public function deleteOptionAction(Request $request)
     {
         $types = [
-            'tickets'       => 'CustomDefTicket',
-            'organizations' => 'CustomDefOrganization',
-            'people'        => 'CustomDefPerson',
-            'chats'         => 'CustomDefChat',
-            'kb'            => 'CustomDefArticle',
-            'download'      => 'CustomDefDownload',
+            'tickets'       => CustomDefTicket::class,
+            'organizations' => CustomDefOrganization::class,
+            'people'        => CustomDefPerson::class,
+            'chats'         => CustomDefChat::class,
+            'kb'            => CustomDefArticle::class,
+            'download'      => CustomDefDownload::class,
         ];
 
         if (!$repClass = @$types[$this->in->getString('type')]) {
@@ -300,22 +306,25 @@ class CustomFieldsController extends AbstractController implements ProtectedCont
             throw new BadRequestHttpException();
         }
         /** @var CustomDefAbstract $rep */
-        $rep  = $this->em->getRepository('DeskPRO:'.$repClass);
+        $rep  = $this->em->getRepository($repClass);
         $step = (int) $this->in->getInt('step');
         switch ($step) {
 
             case 1:
                 $hasData  = $rep->hasData($ids);
-                $response = ['success' => $hasData];
+                $response = ['success' => true];
                 $root     = (int) min($ids);
 
-                if (!$hasData) {
-                    $rep->delete($ids);
+                $field = $rep->getByOptions($ids);
 
-                    return $this->createJsonResponse($response);
+                if (!$hasData) {
+                    if ($repClass !== CustomDefTicket::class || count($rep->getOptionUsage($field, $ids)) === 0) {
+                        $rep->delete($ids);
+
+                        return $this->createJsonResponse(['success' => false]);
+                    }
                 }
 
-                $field   = $rep->getByOptions($ids);
                 $options = [];
                 $map     = [];
                 foreach ($field->children as $child) {
@@ -343,7 +352,12 @@ class CustomFieldsController extends AbstractController implements ProtectedCont
                 if (!$to = $this->in->getInt('update_to')) {
                     throw new BadRequestHttpException();
                 }
-                $rep->updateTo($ids, $to);
+                $field = null;
+                if ($repClass === CustomDefTicket::class) {
+                    $field = $rep->getByOptions($ids);
+                }
+                $rep->updateTo($ids, $to, $field);
+                $rep->delete($ids);
 
                 return $this->createSuccessResponse();
                 break;
